pyproject.toml
[tool.poetry]
name = "seng"
version = "0.1.0"
description = ""
authors = ["Eliot"]
license = "GNU Affero General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
